Client side bundle loading with workarounds for (most) browser bugs
This uses the approach from https://medium.com/@WebReflection/a-universal-bundle-loader-6d7f3e628f93
Basically the deferring of 'ubl.js' should ensure browsers with bugs will load that file, but not any more files if ESM is supported.
I did simplify the script slightly, to only have a ES2017 and a ES5 bundle (not a ES2015 variant).